requirements.txt
pyramid==1.6.1
docker-py==1.7.2

